refactor: (fe) 멘션 기능 파일 분리 및 리팩터링 (#529)
