+++ frontend/src/components/CommentInput/useCommentInput.ts
@@ -1,370 +1,27 @@
 import { UseCommentInputProps } from './type';
 import { queryKeys } from 'apis/constants';
-import {
-  useGetMembers,
-  usePatchComments,
-  usePostComment,
-  usePostMentionNotifications,
-} from 'apis/feedApi';
-import { useState, useEffect, useRef, KeyboardEventHandler } from 'react';
+import { usePatchComments, usePostComment } from 'apis/feedApi';
+import { useState, useEffect, useRef } from 'react';
 import { useQueryClient } from 'react-query';
 import { useParams } from 'react-router-dom';
 import { useRecoilValue } from 'recoil';
 import { isLoginState } from 'recoil/auth/atoms';
-import { getCursorPosition, insertAfter } from 'utils';
 
-import useMutationObserver from 'hooks/useMutationObserver';
 import useSnackBar from 'hooks/useSnackBar';
+
+import useMention from 'components/CommentInput/useMention';
 
 import { MAX_TEXTAREA_LENGTH } from 'constants/domain';
 import { CLIENT_PATH } from 'constants/path';
 
-const DEFAULT_INPUT_HEIGHT = '1.5rem';
 const INITIAL_CONTENT = '';
-const ABSENCE_SYMBOL_POSITION = -1;
+const DEFAULT_INPUT_HEIGHT = '1.5rem';
 
 const useCommentInput = ({
   selectedCommentId,
   editMode,
   turnOffEditMode,
 }: UseCommentInputProps) => {
-  // ------- 멘션 알림 기능 ------------------------
-  const isFirstRendered = useRef(true);
-  const [filterValue, setFilterValue] = useState('');
-  const [mentionedMemberIds, setMentionedMemberIds] = useState<Array<number>>([]);
-  const lastMentionSymbolPositionRef = useRef(ABSENCE_SYMBOL_POSITION);
-  const isFilterValueInitiated = useRef(false);
-  const isPrevPressBackspace = useRef(false);
-  const prevCursorPosition = useRef(0);
-
-  const {
-    isFetching: isFetchingMembers,
-    data: membersData,
-    hasNextPage: hasNextMembersPage,
-    fetchNextPage: fetchNextMembersPage,
-    refetch: refetchMembers,
-  } = useGetMembers(
-    { filter: filterValue },
-    {
-      onSuccess: (data) => {
-        if (data.pages[0].data.length === 0) {
-          handleClosePopover();
-          return;
-        }
-
-        handleOpenPopover();
-      },
-      enabled: false,
-      suspense: false,
-      staleTime: 600000, // 10분
-    },
-  );
-
-  const {
-    mutate: postMentionNotifications,
-    isLoading: isLoadingPostMentionNotifications,
-  } = usePostMentionNotifications();
-
-  useEffect(() => {
-    if (isFirstRendered.current) {
-      isFirstRendered.current = false;
-      return;
-    }
-    if (isFilterValueInitiated.current === true) {
-      return;
-    }
-
-    refetchMembers();
-  }, [filterValue]);
-
-  useEffect(() => {
-    const handleKeydown: KeyboardEventHandler<HTMLDivElement> = (event) => {
-      if (event.key === 'Backspace' || event.key === 'Delete') {
-        // refetch
-        detectMentionSymbolWhenTextDeleted();
-        return;
-      }
-
-      isPrevPressBackspace.current = false;
-      prevCursorPosition.current = 0;
-
-      if (event.key === 'ArrowLeft') {
-        // closePopover
-        detectLeftEscapingMentionArea();
-        return;
-      }
-
-      if (event.key === 'ArrowRight') {
-        detectRightEscapingMentionArea();
-        return;
-      }
-    };
-    commentInputRef.current?.addEventListener('keydown', handleKeydown);
-
-    return () => {
-      commentInputRef.current?.removeEventListener('keydown', handleKeydown);
-    };
-  }, []);
-
-  const inputChangeHandler: MutationCallback = (mutations) => {
-    mutations.forEach((mutation) => {
-      if (mutation.type === 'childList') {
-        if (
-          mutation.removedNodes.length > 0 &&
-          mutation.removedNodes[0].nodeName === 'SPAN'
-        ) {
-          const targetNode = mutation.removedNodes[0] as HTMLElement;
-          const deletedMemberId = Number(targetNode.getAttribute('data-member-id'));
-          const deletedIndex = mentionedMemberIds.indexOf(deletedMemberId);
-          if (deletedIndex > -1) {
-            mentionedMemberIds.splice(deletedIndex, 1);
-          }
-
-          setMentionedMemberIds((prevMentionMemberIds) => {
-            const copiedMentionMembersId = [...prevMentionMemberIds];
-
-            const deletedIndex = copiedMentionMembersId.indexOf(deletedMemberId);
-
-            if (deletedIndex > -1) {
-              copiedMentionMembersId.splice(deletedIndex, 1);
-            }
-
-            return copiedMentionMembersId;
-          });
-        }
-      }
-    });
-    const hasSymbolPosition =
-      lastMentionSymbolPositionRef.current !== ABSENCE_SYMBOL_POSITION;
-
-    const isCurrentCharacterWhiteSpace = (text: string) =>
-      text[getCursorPosition(commentInputRef.current!)! - 1] === ' ';
-
-    if (isFilterValueInitiated.current === true) {
-      isFilterValueInitiated.current = false;
-    }
-
-    const setNicknameAfterMentionSymbol = (text: string) => {
-      if (isCurrentCharacterWhiteSpace(text)) {
-        lastMentionSymbolPositionRef.current = ABSENCE_SYMBOL_POSITION;
-        setFilterValue(''); // 초기화
-        isFilterValueInitiated.current = true;
-
-        handleClosePopover();
-      } else {
-        // 건들지 마시오
-        setFilterValue(
-          text.slice(
-            lastMentionSymbolPositionRef.current,
-            getCursorPosition(commentInputRef.current!)!,
-          ),
-        );
-      }
-    };
-
-    const commentInputElement = commentInputRef.current!;
-    resizeHeight(commentInputElement);
-    const { innerText } = commentInputElement;
-
-    if (hasSymbolPosition) {
-      setNicknameAfterMentionSymbol(innerText);
-
-      if (getCursorPosition(commentInputRef.current!)! === 0) {
-        handleClosePopover();
-      }
-    } else {
-      detectMentionSymbolWhenTextAdded(innerText);
-    }
-
-    setContent(innerText.slice(0, MAX_TEXTAREA_LENGTH));
-  };
-
-  const commentInputRef = useMutationObserver<HTMLDivElement>(inputChangeHandler);
-
-  // 문자열 새로 입력됐을 때
-  const detectMentionSymbolWhenTextAdded = (text: string) => {
-    const cursorPosition = getCursorPosition(commentInputRef.current!)!;
-    const currentCharacter = text[cursorPosition - 1];
-
-    // 1. 현재 cursor 포지션의 바로 앞이 @인 경우에만 @ 이벤트를 호출한다.
-    if (currentCharacter !== '@') {
-      handleClosePopover();
-      return;
-    }
-
-    // 2. 그런데 호출하려는 @의 앞이 공백이 아니면 호출하지 않는다.
-    if (cursorPosition !== 1 && text.length >= 2 && text[cursorPosition - 2] !== ' ') {
-      return;
-    }
-
-    lastMentionSymbolPositionRef.current = cursorPosition;
-    refetchMembers();
-    handleOpenPopover();
-  };
-
-  const detectMentionSymbolWhenTextDeleted = () => {
-    const cursorPosition = getCursorPosition(commentInputRef.current!)!;
-
-    const { innerText } = commentInputRef.current!;
-
-    // 지우려는 element 첫 번째 자식인 경우 contentEditable 부모에서 해당 element를 삭제할 수 없는 이슈 때문에
-    // 다음과 같은 분기문을 통해 첫 번째 element도 삭제 가능하도록 이슈 해결함
-    if (
-      commentInputRef.current?.childNodes.length === 3 &&
-      commentInputRef.current?.childNodes[1].nodeName &&
-      commentInputRef.current?.childNodes[0].textContent === ''
-    ) {
-      // init 3종 세트
-      lastMentionSymbolPositionRef.current = ABSENCE_SYMBOL_POSITION;
-      isFilterValueInitiated.current = true;
-      setFilterValue('');
-      handleClosePopover();
-
-      commentInputRef.current!.textContent = ' ';
-
-      return;
-    }
-
-    if (!innerText.includes('@')) {
-      lastMentionSymbolPositionRef.current = ABSENCE_SYMBOL_POSITION;
-
-      return;
-    }
-
-    const mentionSymbolPosition = innerText.lastIndexOf('@', cursorPosition);
-
-    const targetText = innerText.slice(mentionSymbolPosition + 1, cursorPosition - 1);
-    // 슬라이스 한 문자열 내부에 공백이 있나
-    if (cursorPosition - 1 === mentionSymbolPosition) {
-      lastMentionSymbolPositionRef.current = ABSENCE_SYMBOL_POSITION;
-      return;
-    }
-
-    if (targetText.includes(' ')) {
-      lastMentionSymbolPositionRef.current = ABSENCE_SYMBOL_POSITION;
-      return;
-    }
-
-    // mentionSymbolPosition 앞에 공백이 있나 && mentionSymbolPosition 위치가 첫번째이다.
-    if (mentionSymbolPosition !== 0 && innerText[mentionSymbolPosition - 1] !== ' ') {
-      lastMentionSymbolPositionRef.current = ABSENCE_SYMBOL_POSITION;
-      return;
-    }
-
-    lastMentionSymbolPositionRef.current = mentionSymbolPosition + 1;
-    isPrevPressBackspace.current = true;
-    prevCursorPosition.current = cursorPosition;
-  };
-
-  const detectLeftEscapingMentionArea = () => {
-    const cursorPosition = getCursorPosition(commentInputRef.current!)!;
-    const { innerText } = commentInputRef.current!;
-
-    if (innerText[cursorPosition - 2] === ' ') {
-      lastMentionSymbolPositionRef.current = ABSENCE_SYMBOL_POSITION;
-      setFilterValue('');
-      isFilterValueInitiated.current = true;
-      handleClosePopover();
-    }
-  };
-
-  const detectRightEscapingMentionArea = () => {
-    const cursorPosition = getCursorPosition(commentInputRef.current!)!;
-    const { innerText } = commentInputRef.current!;
-
-    if (innerText[cursorPosition] === ' ') {
-      // init 3종 세트
-      lastMentionSymbolPositionRef.current = ABSENCE_SYMBOL_POSITION;
-      setFilterValue('');
-      isFilterValueInitiated.current = true;
-      handleClosePopover();
-    }
-  };
-
-  const [isPopoverOpen, setIsPopoverOpen] = useState(false);
-  const handleClosePopover = () => {
-    setIsPopoverOpen(false);
-  };
-  const handleOpenPopover = () => {
-    setIsPopoverOpen(true);
-  };
-
-  const selectMember = (memberId: number, nickname: string) => {
-    if (commentInputRef.current === null) {
-      return;
-    }
-
-    mentionedMemberIds.push(memberId);
-    setMentionedMemberIds(mentionedMemberIds);
-
-    const cursorPosition = getCursorPosition(commentInputRef.current)!;
-    const childNodes = commentInputRef.current.childNodes;
-
-    let accLength = 0;
-
-    const currentNodeIndex = Array.from(childNodes).findIndex((childNode) => {
-      const text = childNode.textContent!;
-      if (lastMentionSymbolPositionRef.current <= accLength + text.length) {
-        return true;
-      }
-
-      accLength += text.length;
-      return false;
-    });
-
-    const currentNode = childNodes[currentNodeIndex];
-    const currentText = currentNode.textContent;
-
-    const targetStart = lastMentionSymbolPositionRef.current - accLength - 1;
-    const targetEnd =
-      lastMentionSymbolPositionRef.current - accLength + filterValue.length;
-
-    const frontTextNode = document.createTextNode(
-      currentText!.substring(0, targetStart)!,
-    );
-
-    const nicknameSpan = document.createElement('span');
-    nicknameSpan.innerText = `@${nickname}`;
-    nicknameSpan.setAttribute('contenteditable', 'false');
-    nicknameSpan.style.backgroundColor = '#7b61fe';
-    nicknameSpan.style.color = '#fff';
-    nicknameSpan.style.borderRadius = '5px';
-    nicknameSpan.style.lineHeight = '2';
-    nicknameSpan.style.padding = '2px 4px';
-    nicknameSpan.setAttribute('data-member-id', String(memberId));
-
-    const backTextNode = document.createTextNode(currentText!.substring(targetEnd)!);
-
-    if (currentNodeIndex === 0) {
-      commentInputRef.current.appendChild(frontTextNode);
-      commentInputRef.current.appendChild(nicknameSpan);
-      commentInputRef.current.appendChild(backTextNode);
-    } else {
-      const prevNode = childNodes[currentNodeIndex - 1];
-
-      insertAfter(prevNode, frontTextNode);
-      insertAfter(frontTextNode, nicknameSpan);
-      insertAfter(nicknameSpan, backTextNode);
-    }
-
-    commentInputRef.current.removeChild(currentNode);
-
-    // init 삼형제
-    lastMentionSymbolPositionRef.current = ABSENCE_SYMBOL_POSITION;
-    setFilterValue('');
-    isFilterValueInitiated.current = true;
-
-    // contenteditable div에서 cursor position에 focus 하기
-    const element = commentInputRef.current;
-    const selection = window.getSelection();
-    const range = document.createRange();
-    selection!.removeAllRanges();
-    range.selectNodeContents(element);
-    range.collapse(false);
-    selection!.addRange(range);
-    element.focus();
-  };
-
   // ---------- 댓글 작성 및 수정하여 db에 보내는 관련 로직 ----------------
 
   const [content, setContent] = useState(INITIAL_CONTENT);
@@ -372,6 +29,25 @@ const useCommentInput = ({
   const isLogin = useRecoilValue(isLoginState);
   const { cycleDetailId } = useParams();
   const renderSnackBar = useSnackBar();
+
+  const commentInputRef = useRef<HTMLDivElement>(null);
+  const [mentionedMemberIds, setMentionedMemberIds] = useState<Array<number>>([]);
+
+  const {
+    postMentionNotifications,
+    isLoadingPostMentionNotifications,
+    isPopoverOpen,
+    handleClosePopover,
+    membersData,
+    hasNextMembersPage,
+    fetchNextMembersPage,
+    selectMember,
+  } = useMention({
+    commentInputRef,
+    setContent,
+    mentionedMemberIds,
+    setMentionedMemberIds,
+  });
 
   const { mutate: postComment, isLoading: isLoadingPostComment } = usePostComment(
     { cycleDetailId: Number(cycleDetailId) },
@@ -418,7 +94,7 @@ const useCommentInput = ({
 
   useEffect(() => {
     setContent(editMode.editContent);
-    commentInputRef!.current!.innerText = editMode.editContent;
+    commentInputRef.current!.innerText = editMode.editContent;
   }, [editMode]);
 
   const isVisibleWriteButton = content.length !== 0;
@@ -453,12 +129,6 @@ const useCommentInput = ({
     queryClient.invalidateQueries(queryKeys.getCommentsById);
   };
 
-  const resizeHeight = (element: HTMLDivElement) => {
-    // 입력 값을 지웠을 때, 댓글 입력창의 높이를 줄이기 위한 코드
-    element.style.height = DEFAULT_INPUT_HEIGHT;
-    element.style.height = element.scrollHeight + 'px';
-  };
-
   const resizeToInitialHeight = () => {
     if (!commentInputRef.current) {
       return;
@@ -475,14 +145,13 @@ const useCommentInput = ({
     isLoadingPostComment,
     isLoadingPatchComment,
     handleClickWrite,
-    isFetchingMembers,
+    isLoadingPostMentionNotifications,
+    isPopoverOpen,
+    handleClosePopover,
     membersData,
     hasNextMembersPage,
     fetchNextMembersPage,
-    isPopoverOpen,
-    handleClosePopover,
     selectMember,
-    isLoadingPostMentionNotifications,
   };
 };
